+++ src/main/php/ioc/IoBindingModule.php
@@ -20,30 +20,39 @@ use stubbles\webapp\websession\SessionBindingScope;
 class IoBindingModule implements BindingModule
 {
     /**
+     * marker whether runtime was already initialized
+     *
+     * @type  bool
+     */
+    private static $initialized = false;
+
+    /**
+     * checks whether runtime was already bound
+     *
+     * @internal
+     * @return  bool
+     */
+    public static function initialized()
+    {
+        return self::$initialized;
+    }
+
+    /**
+     * resets initialzed status
+     *
+     * @internal
+     */
+    public static function reset()
+    {
+        self::$initialized = false;
+    }
+
+    /**
      * response class to be used
      *
      * @type  string
      */
     private $responseClass  = 'stubbles\webapp\response\WebResponse';
-    /**
-     * map of formatters for mime types
-     *
-     * @type  array
-     */
-    private $formatter      = ['application/json'    => 'stubbles\webapp\response\format\JsonFormatter',
-                               'text/json'           => 'stubbles\webapp\response\format\JsonFormatter',
-                               'text/html'           => 'stubbles\webapp\response\format\HtmlFormatter',
-                               'text/plain'          => 'stubbles\webapp\response\format\PlainTextFormatter'
-                              ];
-    /**
-     * map of xml formatters for mime types
-     *
-     * @var  array
-     */
-    private $xmlFormatter   = ['text/xml'            => 'stubbles\webapp\response\format\XmlFormatter',
-                               'application/xml'     => 'stubbles\webapp\response\format\XmlFormatter',
-                               'application/rss+xml' => 'stubbles\webapp\response\format\XmlFormatter'
-                              ];
     /**
      * function that creates the session instance
      *
@@ -85,19 +94,6 @@ class IoBindingModule implements BindingModule
     }
 
     /**
-     * adds formatter class for given mime type
-     *
-     * @param   string  $mimeType   mime type that should be handled by given formatter class
-     * @param   string  $formatter  class to handle given mime type
-     * @return  \stubbles\webapp\ioc\IoBindingModule
-     */
-    public function addFormatter($mimeType, $formatter)
-    {
-        $this->formatter[$mimeType] = $formatter;
-        return $this;
-    }
-
-    /**
      * configure the binder
      *
      * @param  \stubbles\ioc\Binder  $binder
@@ -113,37 +109,11 @@ class IoBindingModule implements BindingModule
                ->toInstance($request);
         $binder->bind('stubbles\webapp\response\Response')
                ->toInstance($response);
-        $formatters = $this->availableFormatters();
-        $binder->bindConstant('stubbles.webapp.response.format.mimetypes')
-               ->to(array_keys($formatters));
-        foreach ($formatters as $mimeType => $formatter) {
-            $binder->bind('stubbles\webapp\response\format\Formatter')
-                   ->named($mimeType)
-                   ->to($formatter);
-        }
-
         if (null !== $this->sessionCreator) {
             $sessionCreator = $this->sessionCreator;
             $session = $sessionCreator($request, $response);
             $binder->bind('stubbles\webapp\session\Session')->toInstance($session);
             $binder->setSessionScope(new SessionBindingScope($session));
         }
-    }
-
-    /**
-     * returns map of available formatters
-     *
-     * @return  array
-     */
-    private function availableFormatters()
-    {
-        $formatter = $this->formatter;
-        foreach ($this->xmlFormatter as $mimeType => $xmlFormatter) {
-            if (!isset($formatter[$mimeType]) && class_exists('stubbles\xml\serializer\XmlSerializerFacade')) {
-                $formatter[$mimeType] = $xmlFormatter;
-            }
-        }
-
-        return $formatter;
     }
 }